Updated chart colours, made more consistent on reloads
